chore: move all extension client code to extension folder
