+++ extension/package.json
