+++ extension/src/manifest.json
