+++ extension/src/scripts/content.js
